refactor(wv-match): extract transfer endpoints + helpers (sonarjs/cognitive-complexity, max-lines)
acceptWithTransfer (19) bundled IDOR validation, the split-vs-direct
divisor mutation, target re-assignment, and donor status recomputation,
keeping wvImportMatchController.ts at 1040 lines (over the 1000 limit).

Moved acceptWithTransfer, getTransferPreview, and three new helpers
(verifyTransferOwnership, executeTransferStep,
recomputeDonorMatchStatus) to wvImportMatchTransfer.ts; the main
controller re-exports them so adminRoutes / worldViewImportController
imports stay unchanged.
